Revert "MNT: refactor parachute implementation (#958)"
This reverts commit b516ace01d55821c2fe2395674a8d10052035d2e.
