+++ docs/user/deployable.rst
@@ -8,7 +8,7 @@ Let's start by importing the rocketpy classes we will use.
 
 .. jupyter-execute::
 
-      from rocketpy import Environment, SolidMotor, Rocket, Flight, HemisphericalParachute
+      from rocketpy import Environment, SolidMotor, Rocket, Flight
 
 
 Creating Environment
@@ -182,7 +182,7 @@ Therefore we should be careful with the value of its mass.
 
 
       # Define Parachutes for the rocket
-      main_chute = HemisphericalParachute(
+      main_chute = rocket_without_payload.add_parachute(
           "Main",
           cd_s=7.2,
           trigger=800,
@@ -191,7 +191,7 @@ Therefore we should be careful with the value of its mass.
           noise=(0, 8.3, 0.5),
       )
 
-      drogue_chute = HemisphericalParachute(
+      drogue_chute = rocket_without_payload.add_parachute(
           "Drogue",
           cd_s=0.72,
           trigger="apogee",
@@ -199,8 +199,6 @@ Therefore we should be careful with the value of its mass.
           lag=1.5,
           noise=(0, 8.3, 0.5),
       )
-      rocket_without_payload.add_parachute(parachute = main_chute)
-      rocket_without_payload.add_parachute(parachute = drogue_chute)
 
 .. jupyter-execute::
 
@@ -244,7 +242,7 @@ surfaces to stabilize it, nor a motor that ignites. It does, however, have parac
           center_of_mass_without_motor=0,
       )
 
-      payload_drogue = HemisphericalParachute(
+      payload_drogue = payload_rocket.add_parachute(
           "Drogue",
           cd_s=0.35,
           trigger="apogee",
@@ -253,7 +251,7 @@ surfaces to stabilize it, nor a motor that ignites. It does, however, have parac
           noise=(0, 8.3, 0.5),
       )
 
-      payload_main = HemisphericalParachute(
+      payload_main = payload_rocket.add_parachute(
           "Main",
           cd_s=4.0,
           trigger=800,
@@ -261,8 +259,6 @@ surfaces to stabilize it, nor a motor that ignites. It does, however, have parac
           lag=1.5,
           noise=(0, 8.3, 0.5),
       )
-      payload_rocket.add_parachute(parachute = payload_drogue)
-      payload_rocket.add_parachute(parachute = payload_main)
 
 .. important::
