+++ docs/user/first_simulation.rst
@@ -67,7 +67,7 @@ we will use from RocketPy:
 
 .. jupyter-execute::
 
-    from rocketpy import Environment, SolidMotor, Rocket, Flight, HemisphericalParachute
+    from rocketpy import Environment, SolidMotor, Rocket, Flight
 
 .. note::
 
@@ -272,7 +272,7 @@ Finally, we can add any number of Parachutes to the ``Rocket`` object.
 
 .. jupyter-execute::
 
-    main = HemisphericalParachute(
+    main = calisto.add_parachute(
         name="main",
         cd_s=10.0,
         trigger=800,      # ejection altitude in meters
@@ -284,7 +284,7 @@ Finally, we can add any number of Parachutes to the ``Rocket`` object.
         porosity=0.0432,
     )
 
-    drogue = HemisphericalParachute(
+    drogue = calisto.add_parachute(
         name="drogue",
         cd_s=1.0,
         trigger="apogee",  # ejection at apogee
@@ -295,8 +295,6 @@ Finally, we can add any number of Parachutes to the ``Rocket`` object.
         height=1.5,
         porosity=0.0432,
     )
-    calisto.add_parachute(parachute = main)
-    calisto.add_parachute(parachute = drogue)
 
 We can then see if the rocket is stable by plotting the static margin:
