--- rocketpy/rocket/parachutes/__init__.py
@@ -1,2 +1,0 @@
-from .hemispherical_parachute import HemisphericalParachute
-from .parachute import Parachute
