+++ tests/fixtures/rockets/rocket_fixtures.py
@@ -1,7 +1,7 @@
 import numpy as np
 import pytest
 
-from rocketpy import HemisphericalParachute, Rocket
+from rocketpy import Rocket
 
 
 @pytest.fixture
@@ -404,18 +404,14 @@ def prometheus_rocket(generic_motor_cesaroni_M1520):
         position=0.273,
         sweep_length=0.066,
     )
-    drogue = HemisphericalParachute(
+    prometheus.add_parachute(
         "Drogue",
         cd_s=1.6 * np.pi * 0.3048**2,  # Cd = 1.6, D_chute = 24 in
         trigger="apogee",
-        sampling_rate=100,
     )
-    main = HemisphericalParachute(
+    prometheus.add_parachute(
         "Main",
         cd_s=2.2 * np.pi * 0.9144**2,  # Cd = 2.2, D_chute = 72 in
         trigger=457.2,  # 1500 ft
-        sampling_rate=100,
     )
-    prometheus.add_parachute(parachute=drogue)
-    prometheus.add_parachute(parachute=main)
     return prometheus
